[Canvas] `component/`-level separation of `workpad_page` (#34983)
Refactor: structure workpad_page into smaller static and interactive parts
# Conflicts:
#	x-pack/plugins/canvas/public/components/workpad_page/workpad_interactive_page/interactive_workpad_page.js
#	x-pack/plugins/canvas/public/components/workpad_page/workpad_static_page/static_workpad_page.js
