+++ x-pack/plugins/canvas/public/components/workpad_page/workpad_interactive_page/event_handlers.js
