+++ x-pack/plugins/canvas/public/components/workpad_page/workpad_interactive_page/workpad_interactive_page.scss
@@ -1,0 +1,3 @@
+.canvasInteractivePage {
+  user-select: none;
+}
