+++ x-pack/plugins/canvas/public/components/workpad_page/workpad_static_page/workpad_static_page.scss
@@ -1,0 +1,4 @@
+.canvasLayout__stage .canvasStaticPage.isInactive {
+  visibility: hidden;
+  opacity: 0;
+}
