+++ x-pack/plugins/canvas/public/style/index.scss
@@ -54,6 +54,8 @@
 @import '../components/workpad_loader/workpad_loader';
 @import '../components/workpad_loader/workpad_dropzone/workpad_dropzone';
 @import '../components/workpad_page/workpad_page';
+@import '../components/workpad_page/workpad_interactive_page/workpad_interactive_page';
+@import '../components/workpad_page/workpad_static_page/workpad_static_page';
 
 @import '../../canvas_plugin_src/renderers/advanced_filter/component/advanced_filter.scss';
 @import '../../canvas_plugin_src/renderers/dropdown_filter/component/dropdown_filter.scss';
